requirements.txt
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
cffi==1.14.4
coverage==5.4
cryptography==3.3.1
decorator==4.4.2
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jeepney==0.6.0
keyring==21.8.0
ledger-autosync==1.0.2
lxml==4.6.2
ofxclient==2.0.3
ofxhome==0.3.3
ofxparse==0.20
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.14
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
PyYAML==5.4.1
SecretStorage==3.3.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
